后台代码很好写，SpuBo extends Spu， 主要是前台要组织好数据。

+++ ly-item/ly-item-interface/src/main/java/com/leyou/item/pojo/Sku.java
@@ -31,6 +31,6 @@ public class Sku {
     private Date lastUpdateTime;
 
     @Transient
-    private Long stock;
+    private Integer stock;
 
 }
